feat: adding useToggleButtonGroup hook
